+++ ingress/tools.go
@@ -96,7 +96,7 @@ func (h *Handler) GetAge(object interface{}) (time.Duration, error) {
 	case string:
 		ing, err := h.Get(val)
 		if err != nil {
-			return time.Duration(int64(0)), err
+			return time.Duration(0), err
 		}
 		return time.Now().Sub(ing.CreationTimestamp.Time), nil
 	case *networkingv1.Ingress:
@@ -104,6 +104,6 @@ func (h *Handler) GetAge(object interface{}) (time.Duration, error) {
 	case networkingv1.Ingress:
 		return time.Now().Sub(val.CreationTimestamp.Time), nil
 	default:
-		return time.Duration(int64(0)), ErrInvalidToolsType
+		return time.Duration(0), ErrInvalidToolsType
 	}
 }
