+++ secret/tools.go
@@ -48,7 +48,7 @@ func (h *Handler) GetAge(object interface{}) (time.Duration, error) {
 	case string:
 		secret, err := h.Get(val)
 		if err != nil {
-			return time.Duration(int64(0)), err
+			return time.Duration(0), err
 		}
 		return time.Now().Sub(secret.CreationTimestamp.Time), nil
 	case *corev1.Secret:
@@ -56,6 +56,6 @@ func (h *Handler) GetAge(object interface{}) (time.Duration, error) {
 	case corev1.Secret:
 		return time.Now().Sub(val.CreationTimestamp.Time), nil
 	default:
-		return time.Duration(int64(0)), ErrInvalidToolsType
+		return time.Duration(0), ErrInvalidToolsType
 	}
 }
